scripts/instance-requirements.txt
boto3
python-secrets
requests 
configparser
simplejson 
python-dateutil 
gophish
python-docx
user-agents
click
xlsxwriter
